% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_helpers.R
\name{all_duplicated}
\alias{all_duplicated}
\title{Identify all duplicates}
\usage{
all_duplicated(x, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}

\item{...}{arguments for particular methods.}
}
\description{
Identifies all duplicates, see \code{\link[base:duplicated]{base::duplicated()}} for details.
}
