% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_df.R
\name{describe_cols}
\alias{describe_cols}
\title{Describe the columns in a data frame}
\usage{
describe_cols(df, ...)
}
\arguments{
\item{df}{a data.frame}

\item{...}{additional arguments passed on to \code{\link[=map_unique]{map_unique()}}}
}
\description{
For each column in a data.frame, \code{describe_cols()} shows the column type,
number and percentage of missing values, and number of unique values
(optionally with the values themselves--see \code{\link[=map_unique]{map_unique()}} for details).
}
