% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{file_dir_exist}
\alias{file_dir_exist}
\title{Confirms Existence of and Separates Files and Directories}
\usage{
file_dir_exist(paths, if_nonexistent = "capture")
}
\arguments{
\item{paths}{character vectors containing file or directory paths.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is done.}

\item{if_nonexistent}{how to handle paths which do not exist; one of:
"capture" = capture nonexistent paths in a named vector in the list
"warn" = emit a warning listing paths that do not exist
"error" = emit an error listing paths that do not exist}
}
\value{
A 2- or 3-vector, named list with:
- "file" containing path to files,
- "dir" containing path to directories, and
- "nonexistent" containing paths that do not exist, if
\code{if_nonexistent = "capture"}; otherwise paths that do not exist are
listed in a warning or error message
}
\description{
Confirms all paths exist and separates files from directories
}
