% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list.R
\name{filter_list}
\alias{filter_list}
\title{Filter list by name/value}
\usage{
filter_list(list, name, value)
}
\arguments{
\item{list}{a list}

\item{name}{a name to search for within a list; can be anything that can
appear as a name in a list}

\item{value}{a value to search for within a list}
}
\description{
Filters list (at top-level only) by name or value.
}
\section{NOTE}{

\code{filter_list()} is useful for returning matches of a name or value.
Currently, it will only search the top-level in a list and matches must be
exact. Unlike \code{\link[purrr:pluck]{purrr::pluck()}}, \code{filter_list()} will return all matches (not
just the first match).
}

\examples{
l <- list(this = 1, that = 2, that = 3, final = 2)

purrr::pluck(l, "that") # returns only first match
filter_list(l, "that") # returns all matches

filter_list(l, value = 2)

}
