% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_consensus_path.R
\name{find_consensus_path}
\alias{find_consensus_path}
\title{Identify shared path of subdirectories}
\usage{
find_consensus_path(top_lvl_dir)
}
\arguments{
\item{top_lvl_dir}{the path to the directory to search within}
}
\value{
a file path, beginning with \code{top_lvl_dir}
}
\description{
Identify the recursive path (subdirectories) which most files have in common
within a specified directory.
}
\section{Example Usage}{

Identify the consensus path of an \code{renv} project library when other
unrelated and less file-dense directories are present in a given top-level
directory. (I used this on a HPCC to determine where R files were being
installed to when setting up R with renv and conda.)
}

\section{NOTE}{

This is a first-pass implementation and may be slow. The implementation
identifies all files recursively below the specified directory and returns
the file path shared by the majority of files (currently 80\%).
}

