% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_df.R
\name{identify_identical}
\alias{identify_identical}
\title{Returns columns in a data.frame that are identical (strict)}
\usage{
identify_identical(df)
}
\description{
Returns columns in a data.frame that are identical (strict)
}
