% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_df.R
\name{map_missing}
\alias{map_missing}
\title{Show missing observations for all columns in a data frame}
\usage{
map_missing(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Show missing observations for all columns in a data frame
}
