% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_df.R
\name{map_unique}
\alias{map_unique}
\title{Report unique values for each column of a data.frame}
\usage{
map_unique(df, vals = "as_string", sep = " | ", ignore_NA = TRUE)
}
\arguments{
\item{df}{a data.frame}

\item{vals}{a string describing how to display unique values; one of:
\itemize{
\item "as_string" (default): for values as a concatenated string
(for quick viewing)
\item "as_list": for values as a column of lists
\item "both": for concatenated string AND list columns
\item "none"
}}

\item{sep}{a character string to separate the values; used with "as_string"}

\item{ignore_NA}{logical; whether to include missing (\code{NA}) values in report;
affects both count and values displayed}
}
\description{
For each column in a data.frame, \code{map_unique()} shows the number of unique
values and, optionally, the values themselves.
}
