% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_aes_text_warning.R
\name{suppress_aes_text_warning}
\alias{suppress_aes_text_warning}
\title{Suppress ggplot2 warning when using text aesthetic for plotly}
\usage{
suppress_aes_text_warning(ggplot2_expr)
}
\description{
Suppresses unknown aesthetic warning when providing a \code{text} aesthetic.
Intended for cosmetic purposes when printing as a ggplot object.
Unnecessary, if converting to plotly for use as hovertext with \code{ggplotly()}.
}
\examples{
g <- ggplot(iris) +
    geom_point(aes(x = Sepal.Length, y = Sepal.Width, text = Species))
g

suppress_aes_text_warning(g)

}
